Rotation of intensity data removed from mask editor to RealDataSelector actions.
