+++ GUI/coregui/Views/MaskWidgets/MaskEditorCanvas.h
@@ -54,7 +54,6 @@ public slots:
     void onPresentationTypeRequest(MaskEditorFlags::PresentationType presentationType);
     void onSavePlotRequest();
     void onResetViewRequest();
-    void onRotateDataRequest();
 
 private:
     bool isAxisRangeMatchData() const;
